(1)arithmetic_coding目录下，增加了README
