+++ project-spec/meta-user/recipes-apps/autostart/src/arithmetic_coding/ArithmeticCoder.cpp
@@ -1,7 +1,6 @@
 #include <limits>
 #include <stdexcept>
 #include "ArithmeticCoder.hpp"
-#include "BitIoStream.hpp"
 #include <cstdint>
 #include <cstdlib>
 #include <fstream>
@@ -12,6 +11,66 @@
 #include <stddef.h> 
 using std::uint32_t;
 using std::uint64_t;
+
+BitInputStream::BitInputStream(std::istream &in) :
+	input(in),
+	currentByte(0),
+	numBitsRemaining(0) {}
+	
+	
+int BitInputStream::read() {
+	if (currentByte == std::char_traits<char>::eof())
+		return -1;
+	if (numBitsRemaining == 0) {
+		currentByte = input.get();  // Note: istream.get() returns int, not char
+		if (currentByte == std::char_traits<char>::eof())
+			return -1;
+		if (!(0 <= currentByte && currentByte <= 255))
+			throw std::logic_error("Assertion error");
+		numBitsRemaining = 8;
+	}
+	if (numBitsRemaining <= 0)
+		throw std::logic_error("Assertion error");
+	numBitsRemaining--;
+	return (currentByte >> numBitsRemaining) & 1;
+}
+
+
+int BitInputStream::readNoEof() {
+	int result = read();
+	if (result != -1)
+		return result;
+	else
+		throw std::runtime_error("End of stream");
+}
+
+
+
+BitOutputStream::BitOutputStream(uint8_t* _addr) :
+	currentByte(0),
+	numBitsFilled(0) ,
+	data_addr(_addr) {}
+
+void BitOutputStream::write(int b) {
+	if (b != 0 && b != 1)
+		throw std::domain_error("Argument must be 0 or 1");
+	currentByte = (currentByte << 1) | b;
+	numBitsFilled++;
+	if (numBitsFilled == 8) {
+		// Note: ostream.put() takes char, which may be signed/unsigned
+		if (std::numeric_limits<char>::is_signed)
+			currentByte -= (currentByte >> 7) << 8;
+		data_addr[size++]=static_cast<uint8_t>(currentByte);
+		currentByte = 0;
+		numBitsFilled = 0;
+	}
+}
+
+
+void BitOutputStream::finish() {
+	while (numBitsFilled != 0)
+		write(0);
+}
 
 
 ArithmeticCoderBase::ArithmeticCoderBase(int numBits) {
@@ -208,8 +267,8 @@ extern "C" size_t coding(Gmm gmm,uint16_t* trans_addr,size_t trans_len,uint8_t* 
 
 	//下面两行代码是将字符数组转换为bin文件。
 	// 实际中并不需要下面代码，这里得到bin文件，只是为了解码，从而验证编解码的正确性
-	char outputFile[255]="IO/enc.bin";
-	write_bin(data_addr,size,outputFile);
+	// char outputFile[255]="IO/enc.bin";
+	// write_bin(data_addr,size,outputFile);
 
 	return size;
 }
