+++ project-spec/meta-user/recipes-apps/autostart/src/arithmetic_coding/ArithmeticCoder.hpp
@@ -2,60 +2,50 @@
 
 #include <algorithm>
 #include <cstdint>
-#include "BitIoStream.hpp"
 #include "GmmTable.h"
+#include <istream>
+#include <ostream>
 
-/* 
- * Provides the state and behaviors that arithmetic coding encoders and decoders share.
- */
+
+class BitInputStream final {
+	private: 
+		std::istream &input;
+		int currentByte;
+		int numBitsRemaining; 
+	public: 
+		int read();
+		int readNoEof();
+		explicit BitInputStream(std::istream &in);
+};
+
+class BitOutputStream final {
+	private: 
+		int currentByte;
+		int numBitsFilled;
+	public:
+		uint8_t* data_addr;
+		size_t size=0;
+		BitOutputStream(uint8_t* _addr);	
+	    void write(int b);
+		void finish();
+};
+
+
 class ArithmeticCoderBase {
 	protected: int numStateBits;
-	
-	// Maximum range (high+1-low) during coding (trivial), which is 2^numStateBits = 1000...000.
-	protected: std::uint64_t fullRange;
-	
-	// The top bit at width numStateBits, which is 0100...000.
-	protected: std::uint64_t halfRange;
-	
-	// The second highest bit at width numStateBits, which is 0010...000. This is zero when numStateBits=1.
-	protected: std::uint64_t quarterRange;
-	
-	// Minimum range (high+1-low) during coding (non-trivial), which is 0010...010.
-	protected: std::uint64_t minimumRange;
-	
-	// Maximum allowed total from a frequency table at all times during coding.
-	protected: std::uint64_t maximumTotal;
-	
-	// Bit mask of numStateBits ones, which is 0111...111.
-	protected: std::uint64_t stateMask;
-	
-	
-	/*---- State fields ----*/
-	
-	// Low end of this arithmetic coder's current range. Conceptually has an infinite number of trailing 0s.
-	protected: std::uint64_t low;
-	
-	// High end of this arithmetic coder's current range. Conceptually has an infinite number of trailing 1s.
-	protected: std::uint64_t high;
-	
-	
-	/*---- Constructor ----*/
-	
-	// Constructs an arithmetic coder, which initializes the code range.
-	public: explicit ArithmeticCoderBase(int numBits);
-	
-	
-	public: virtual ~ArithmeticCoderBase() = 0;
-
-	protected: virtual void update(const GmmTable &freqs, std::uint32_t symbol);
-	
-	
-	// Called to handle the situation when the top bit of 'low' and 'high' are equal.
-	protected: virtual void shift() = 0;
-	
-	
-	// Called to handle the situation when low=01(...) and high=10(...).
-	protected: virtual void underflow() = 0;
+		std::uint64_t fullRange;
+		std::uint64_t halfRange;
+		std::uint64_t quarterRange;
+		std::uint64_t minimumRange;
+		std::uint64_t maximumTotal;
+		std::uint64_t stateMask;
+		std::uint64_t low;
+		std::uint64_t high;
+		explicit ArithmeticCoderBase(int numBits);
+		virtual ~ArithmeticCoderBase() = 0;
+		virtual void update(const GmmTable &freqs, std::uint32_t symbol);
+		virtual void shift() = 0;
+		virtual void underflow() = 0;
 	
 };
