--- project-spec/meta-user/recipes-apps/autostart/src/arithmetic_coding/CMakeLists.txt
@@ -1,15 +1,0 @@
-cmake_minimum_required(VERSION 3.10)
-
-project(ArithmeticCoding)
-
-add_executable(test 
-    test.cpp 
-    ArithmeticCoder.cpp 
-    BitIoStream.cpp 
-    GmmTable.cpp
-)
-
-
-
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
