--- project-spec/meta-user/recipes-apps/autostart/src/arithmetic_coding/IO/dec.txt
@@ -1,1 +1,0 @@
-10 41 72 103 134 166 197 228 259 291 322 353 384 416 447 478 509 540 572 603 634 665 697 728 759 790 822 853 884 915 947 978 1009 1040 1071 1103 1134 1165 1196 1228 1259 1290 1321 1353 1384 1415 1446 1478 1509 1540 1571 1602 1634 1665 1696 1727 1759 1790 1821 1852 1884 1915 1946 1977 2009 2040 2071 2102 2133 2165 2196 2227 2258 2290 2321 2352 2383 2415 2446 2477 2508 2539 2571 2602 2633 2664 2696 2727 2758 2789 2821 2852 2883 2914 2946 2977 3008 3039 3070 3102 3133 3164 3195 3227 3258 3289 3320 3352 3383 3414 3445 3477 3508 3539 3570 3601 3633 3664 3695 3726 3758 3789 3820 3851 3883 3914 3945 3976 4008 4039 4070 4101 4132 4164 4195 4226 4257 4289 4320 4351 4382 4414 4445 4476 4507 4538 4570 4601 4632 4663 4695 4726 4757 4788 4820 4851 4882 4913 4945 4976 5007 5038 5069 5101 5132 5163 5194 5226 5257 5288 5319 5351 5382 5413 5444 5476 5507 5538 5569 5600 5632 5663 5694 5725 5757 5788 5819 5850 5882 5913 5944 5975 6007 6038 6069 6100 6131 6163 6194 6225 6256 6288 6319 6350 6381 6413 6444 6475 6506 6537 6569 6600 6631 6662 6694 6725 6756 6787 6819 6850 6881 6912 6944 6975 7006 7037 7068 7100 7131 7162 7193 7225 7256 7287 7318 7350 7381 7412 7443 7475 7506 7537 7568 7599 7631 7662 7693 7724 7756 7787 7818 7849 7881 7912 7943 7974 8006 8037 8068 8099 8130 8162 8193 8224 8255 8287 8318 8349 8380 8412 8443 8474 8505 8536 8568 8599 8630 8661 8693 8724 8755 8786 8818 8849 8880 8911 8943 8974 9005 9036 9067 9099 9130 9161 9192 9224 9255 9286 9317 9349 9380 9411 9442 9474 9505 9536 9567 9598 9630 9661 9692 9723 9755 9786 9817 9848 9880 9911 9942 9973 10005 10036 10067 10098 10129 10161 10192 10223 10254 10286 10317 10348 10379 10411 10442 10473 10504 10535 10567 10598 10629 10660 10692 10723 10754 10785 10817 10848 10879 10910 10942 10973 11004 11035 11066 11098 11129 11160 11191 11223 11254 11285 11316 11348 11379 11410 11441 11473 11504 11535 11566 11597 11629 11660 11691 11722 11754 11785 11816 11847 11879 11910 11941 11972 12004 12035 12066 12097 12128 12160 12191 12222 12253 12285 12316 12347 12378 12410 12441 12472 12503 12534 12566 12597 12628 12659 12691 12722 12753 12784 12816 12847 12878 12909 12941 12972 13003 13034 13065 13097 13128 13159 13190 13222 13253 13284 13315 13347 13378 13409 13440 13472 13503 13534 13565 13596 13628 13659 13690 13721 13753 13784 13815 13846 13878 13909 13940 13971 14003 14034 14065 14096 14127 14159 14190 14221 14252 14284 14315 14346 14377 14409 14440 14471 14502 14533 14565 14596 14627 14658 14690 14721 14752 14783 14815 14846 14877 14908 14940 14971 15002 15033 15064 15096 15127 15158 15189 15221 15252 15283 15314 15346 15377 15408 15439 15471 15502 15533 15564 15595 15627 15658 15689 15720 15752 15783 15814 15845 15877 15908 15939 15970 16002 16033 16064 16095 16126 16158 16189 16220 16251 16283 16314 16345 16376 16408 16439 16470 16501 16532 16564 16595 16626 16657 16689 16720 16751 16782 16814 16845 16876 16907 16939 16970 17001 17032 17063 17095 17126 17157 17188 17220 17251 17282 17313 17345 17376 17407 17438 17470 17501 17532 17563 17594 17626 17657 17688 17719 17751 17782 17813 17844 17876 17907 17938 17969 18001 18032 18063 18094 18125 18157 18188 18219 18250 18282 18313 18344 18375 18407 18438 18469 18500 18531 18563 18594 18625 18656 18688 18719 18750 18781 18813 18844 18875 18906 18938 18969 19000 19031 19062 19094 19125 19156 19187 19219 19250 19281 19312 19344 19375 19406 19437 19469 19500 19531 19562 19593 19625 19656 19687 19718 19750 19781 19812 19843 19875 19906 19937 19968
