+++ project-spec/meta-user/recipes-apps/autostart/src/transmission_protocol.h
@@ -3,7 +3,9 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-#include "coding.h"
+//增加算术编码涉及到的头文件： "编解码器"，"GMM和频率表"
+#include "arithmetic_coding/ArithmeticCoder.hpp"
+#include "arithmetic_coding/GmmTable.h"
 
 __BEGIN_DECLS
